khjkg

+++ src/main/java/frc/robot/commands/Driver/Shoot.java
@@ -25,10 +25,10 @@ public class Shoot extends Command {
 
   private boolean m_isFinished = false;
 
-  private int m_shooterArmPosition = 0;
-  private int m_shooterSpeed = 0;
-  private int m_frontIntakePosition = 0;
-  private int m_frontIntakeSpeed = 0;
+  private double m_shooterArmPosition = 0;
+  private double m_shooterSpeed = 0;
+  private double m_frontIntakePosition = 0;
+  private double m_frontIntakeSpeed = 0;
 
 
   private int m_stateMachine = 1;
